enforcing all lower case standard for package names 'JobScheduler' to 'scheduler' etc
