+++ src/test/java/com/google/gcs/sdrs/scheduler/JobSchedulerTest.java
@@ -16,42 +16,44 @@
  *
  */
 
-package com.google.gcs.sdrs.JobManager;
+package com.google.gcs.sdrs.scheduler;
 
-import com.google.gcs.sdrs.worker.BaseWorker;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
+import com.google.gcs.sdrs.scheduler.JobScheduler;
+
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
-/** Test class for JobManager */
-public class JobManagerTest {
-  private JobManager instance;
+public class JobSchedulerTest {
 
-  /** Set up steps before each test */
+  private JobScheduler instance;
+
+  /**
+   * Set up steps before each test
+   */
   @Before
-  public void setUp() {
-    instance = JobManager.getInstance();
-  }
-
-  /** Tear down steps after each test */
-  @After
-  public void tearDown() {
-    instance.shutDownJobManagerNow();
+  public void setUp(){
+    instance = JobScheduler.getInstance();
   }
 
   /**
-   * Test that a JobManager instance is created when a getInstance is called without an existing
-   * instance
+   * Tear down steps after each test
+   */
+  @After
+  public void tearDown(){
+    instance.shutdownSchedulerNow();
+  }
+
+  /**
+   * Test that a JobManager instance is created when a getInstance is called without an existing instance
    */
   @Test
   public void getInstanceWhenInstanceDoesNotExist() {
     // Instance created in test setup
     assertNotNull(instance);
-    assertNotNull(instance.completionService);
-    assertEquals(instance.activeWorkerCount.get(), 0);
   }
 
   /**
@@ -60,21 +62,7 @@ public class JobManagerTest {
   @Test
   public void getInstanceWhenInstanceAlreadyExists() {
     assertNotNull(instance);
-    JobManager secondInstance = JobManager.getInstance();
+    JobScheduler secondInstance = JobScheduler.getInstance();
     assertEquals(instance, secondInstance);
-  }
-
-  /** Test that the worker count is incremented when a job is submitted */
-  @Test
-  public void testSubmitJob() {
-    assertNotNull(instance);
-    BaseWorker worker =
-        new BaseWorker() {
-          @Override
-          public void doWork() {}
-        };
-    int currentActiveWorkers = instance.activeWorkerCount.get();
-    instance.submitJob(worker);
-    assertEquals(currentActiveWorkers + 1, instance.activeWorkerCount.get());
   }
 }
