+++ src/test/java/com/google/gcs/sdrs/server/ServerShutdownHookTest.java
@@ -23,6 +23,8 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
+import com.google.gcs.sdrs.ServerShutdownHook;
+
 import static org.junit.Assert.assertFalse;
 
 /**
