+++ src/test/java/com/google/gcs/sdrs/service/impl/RetentionRulesServiceImplTest.java
@@ -18,13 +18,13 @@
 package com.google.gcs.sdrs.service.impl;
 
 import com.google.gcs.sdrs.RetentionRuleType;
-import com.google.gcs.sdrs.JobManager.JobManager;
 import com.google.gcs.sdrs.controller.filter.UserInfo;
 import com.google.gcs.sdrs.controller.pojo.RetentionRuleCreateRequest;
 import com.google.gcs.sdrs.controller.pojo.RetentionRuleResponse;
 import com.google.gcs.sdrs.controller.pojo.RetentionRuleUpdateRequest;
 import com.google.gcs.sdrs.dao.impl.RetentionRuleDaoImpl;
 import com.google.gcs.sdrs.dao.model.RetentionRule;
+import com.google.gcs.sdrs.manager.JobManager;
 import com.google.gcs.sdrs.rule.impl.StsRuleExecutor;
 import com.google.gcs.sdrs.worker.Worker;
 import com.google.gcs.sdrs.worker.impl.CancelDefaultJobWorker;
